More restructure and release configuration build settings. - Clean-up of "src/config.h" whereby new header files created and code shuffled around to structure the project better. - Moved version to file and added PROGRAM_VERSION macro setting to Makefile CCFLAGS, in preparation for configurable building for automated CI pipelines. - Similarly also moved Linux Kernel configuration macros to the Makefile. - Building with release configuration or more specifically -O2 option resulted in incorrect register errors when compiling do_div macro code. Suspected the assembler code used to implement do_div is getting impacted by optimization. Disabling optimization for only the few impacted functions resolved the issue; these functions only get called during setup period, bittiming calculations when setting baud-rate. - Renamed QNX resource manager files dev.h/dev.c to resmgr.h/resmgr.c for better clarity and avoid ambiguity with similarly named Linux Kernel files. - Added initial src/kernel/README.md and tests/image/README.md docs.

+++ src/pci.c
@@ -19,11 +19,14 @@
  */
 
 #include <stdio.h>
+#include <string.h>
+
 #include <sys/neutrino.h>
 #include <sys/mman.h>
 
 #include "config.h"
 #include "pci.h"
+#include "main.h"
 
 struct pci_driver *detected_driver = NULL;
 
@@ -288,7 +291,7 @@ int pci_enable_device(struct pci_dev *dev) {
                 dev->ba[i] = ba[i];
 
                 log_info("read ba[%d] { addr: %x, size: %x }\n",
-                        i, ba[i].addr, ba[i].size);
+                        i, (u32)ba[i].addr, (u32)ba[i].size);
             }
         }
 #undef MAX_NUM_BA
@@ -335,7 +338,7 @@ void pci_disable_device(struct pci_dev *dev) {
 }
 
 uintptr_t pci_iomap(struct pci_dev *dev, int bar, unsigned long max) {
-    log_trace("pci_iomap; bar: %d, max: %d\n", bar, max);
+    log_trace("pci_iomap; bar: %d, max: %lu\n", bar, max);
 
     if (bar >= dev->nba) {
         log_err("internal error; bar: %d, nba: %d\n", bar, dev->nba);
@@ -371,7 +374,7 @@ void pci_iounmap(struct pci_dev *dev, uintptr_t p) {
         }
     }
 
-    log_trace("pci_iounmap; bar: %d, size: %d\n", bar, size);
+    log_trace("pci_iounmap; bar: %d, size: %lu\n", bar, size);
 
     if (bar == -1 || !size) {
         log_err("internal error; bar size failure during pci_iounmap\n");
